pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adidt"
version = "0.0.1"
description = "Device tree management tools for ADI hardware"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Analog Devices Inc."}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Eclipse Public License 2.0 (EPL-2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "Click",
    "fdt",
    "fabric",
    "bcrypt==5.0.0",
    "rich",
    "numpy",
    "xmltodict",
    "pydantic>=2",
]

[project.optional-dependencies]
xsa = ["lopper"]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-dotenv",
    "pyadi-iio",
    "pyadi-jif[cplex,draw]",
    "hashable_list",
    "ordered_set",
    "requests",
]
dev = [
    "adidt[test]",
    "adi-labgrid-plugins @ git+https://github.com/tfcollins/labgrid-plugins.git",
    "pyadi-build @ git+https://github.com/tfcollins/pyadi-build.git",
    "usbsdmux",
]
mcp = ["fastmcp"]

#[tool.uv.sources]
#pyadi-build = { path = "./pyadi-build", editable = true }

[project.scripts]
adidtc = "adidt.cli.main:cli"
adidt-mcp = "adidt.mcp_server:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["adidt*"]

[tool.setuptools.package-data]
adidt = ["templates/*.json", "xsa/d3_bundle.js", "xsa/profiles/*.json"]

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# Enforce full Google-style docstrings only in the XSA subpackage.
# All other paths are excluded from docstring rules.
"adidt/xsa/**" = []
"**" = ["D"]

[tool.pytest.ini_options]
env_files = [".env"]
testpaths = ["test"]
norecursedirs = [
    "test/hw",
    "linux",
    "u-boot-xlnx",
    ".nox",
    ".git",
    "build",
    "dist",
    "*.egg",
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --ignore-glob=*_hw.py"
markers = [
    "network: tests requiring internet downloads",
]

[tool.ty.rules]
# Ignore third-party import errors (missing stubs)
unresolved-import = "ignore"

